Safari-611.1.10/LayoutTests/media/modern-media-controls/placard-support/placard-support-error-recover-expected.txt

Testing the PlacardSupport behavior when we recover from an error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.placard is mediaController.controls.invalidPlacard
PASS mediaController.controls.children.includes(mediaController.controls.invalidPlacard) is true
PASS mediaController.controls.placard is null
PASS mediaController.controls.children.includes(mediaController.controls.invalidPlacard) is false

PASS successfullyParsed is true

TEST COMPLETE
